% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlTable_theme.R
\name{getHtmlTableTheme}
\alias{getHtmlTableTheme}
\title{Retrieve the \code{\link{htmlTable}} theme list}
\usage{
getHtmlTableTheme()
}
\value{
\code{list} with the styles to be applied to the table
}
\description{
A wrapper for a \code{\link[base]{getOption}("htmlTable.theme")} call that
returns the standard theme unless one is set.
}
\examples{
getHtmlTableTheme()
}
