% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlTableWidget.R
\name{htmlTableWidget}
\alias{htmlTableWidget}
\title{htmlTable with pagination widget}
\usage{
htmlTableWidget(
  x,
  number_of_entries = c(10, 25, 100),
  width = NULL,
  height = NULL,
  elementId = NULL,
  ...
)
}
\arguments{
\item{x}{A data frame to be rendered}

\item{number_of_entries}{a numeric vector with the number of entries per page to show.
If there is more than one number given, the user will be able to show the number
of rows per page in the table.}

\item{width}{Fixed width for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{height}{Fixed height for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{elementId}{Use an explicit element ID for the widget (rather than an
automatically generated one). Useful if you have other JavaScript that
needs to explicitly discover and interact with a specific widget instance.}

\item{...}{Additional parameters passed to htmlTable}
}
\value{
an htmlwidget showing the paginated table
}
\description{
This widget renders a table with pagination into an htmlwidget
}
