% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyHtmlTable_helpers_innerJoinByCommonCols.r
\name{innerJoinByCommonCols}
\alias{innerJoinByCommonCols}
\title{A simple function for joining two tables by their
intersected columns}
\usage{
innerJoinByCommonCols(x, y)
}
\arguments{
\item{x}{\code{data.frame}}

\item{y}{\code{data.frame}}
}
\value{
\code{data.frame}
}
\description{
A simple function for joining two tables by their
intersected columns
}
