% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{outputInt}
\alias{outputInt}
\title{Deprecated use \code{\link[=txtInt]{txtInt()}} instead.}
\usage{
outputInt(...)
}
\arguments{
\item{...}{Passed to \code{\link[=txtInt]{txtInt()}}}
}
\description{
Deprecated use \code{\link[=txtInt]{txtInt()}} instead.
}
\examples{
\dontrun{
# Deprecated function
outputInt(123456)
}

}
\keyword{internal}
