% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlTable_render_addCells.R
\name{prAddCells}
\alias{prAddCells}
\title{Add a cell}
\usage{
prAddCells(
  rowcells,
  cellcode,
  style_list,
  style,
  prepped_cell_css,
  cgroup_spacer_cells,
  has_rn_col,
  offset = 1,
  style_list_align_key = "align"
)
}
\arguments{
\item{rowcells}{The cells with the values that are to be added}

\item{cellcode}{Type of cell, can either be \code{th} or \code{td}}

\item{style_list}{The style_list}

\item{style}{The cell style}

\item{cgroup_spacer_cells}{The number of cells that occur between
columns due to the cgroup arguments.}

\item{has_rn_col}{Due to the alignment issue we need to keep track
of if there has already been printed a rowname column or not and therefore
we have this has_rn_col that is either 0 or 1.}

\item{offset}{For rgroup rows there may be an offset != 1}
}
\value{
\code{string} Returns the string with the new cell elements
}
\description{
Adds a row of cells \verb{<td>val</td><td>...</td>} to a table string for
\code{\link[=htmlTable]{htmlTable()}}
}
\seealso{
Other hidden helper functions for htmlTable: 
\code{\link{prAddEmptySpacerCell}()},
\code{\link{prAddSemicolon2StrEnd}()},
\code{\link{prEscapeHtml}()},
\code{\link{prGetCgroupHeader}()},
\code{\link{prGetRowlabelPos}()},
\code{\link{prGetStyle}()},
\code{\link{prPrepInputMatrixDimensions}()},
\code{\link{prPrepareAlign}()},
\code{\link{prPrepareCgroup}()},
\code{\link{prTblNo}()}
}
\concept{hidden helper functions for htmlTable}
\keyword{internal}
