% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlTable.R
\name{prConvertDfFactors}
\alias{prConvertDfFactors}
\title{Convert all factors to characters to print them as they expected}
\usage{
prConvertDfFactors(x)
}
\arguments{
\item{x}{The matrix/data.frame with the data. For the \code{print} and \code{knit_print}
it takes a string of the class \code{htmlTable} as \code{x} argument.}
}
\value{
The data frame with factors as characters
}
\description{
Convert all factors to characters to print them as they expected
}
