% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepGroupCounts.R
\name{prepGroupCounts}
\alias{prepGroupCounts}
\title{Retrieves counts for rgroup, cgroup, & tspanner arguments}
\usage{
prepGroupCounts(x)
}
\arguments{
\item{x}{The vector to process}
}
\value{
\code{list(n = rle$lengths, names = rle$values)}
}
\description{
This function is a wrapper to \code{\link[base:rle]{base::rle()}} that
does exactly this but is a little too picky about input values.
}
\examples{
prepGroupCounts(c(1:3, 3:1))
}
