% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlTable_helpers.R
\name{prEscapeHtml}
\alias{prEscapeHtml}
\title{Remove html entities from table}
\usage{
prEscapeHtml(x)
}
\arguments{
\item{x}{}
}
\value{
\code{x} without the html entities
}
\description{
Removes the htmlEntities from table input data. Note that
this also replaces $ signs in order to remove the MathJax
issue.
}
