% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/tidyHtmlTable_helpers_extractElementsAndConvertToTbl.R
\name{prExtractElementsAndConvertToTbl}
\alias{prExtractElementsAndConvertToTbl}
\title{Extract the elements and generate a table with unique elements}
\usage{
prExtractElementsAndConvertToTbl(x, elements)
}
\arguments{
\item{x}{\code{list} with columns to be joined}

\item{elements}{\code{char} vector with the elements to select}
}
\description{
Extract the elements and generate a table with unique elements
}
