% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlTable_helpers_getAlign.R
\name{prGetAlign}
\alias{prGetAlign}
\title{Gets alignment}
\usage{
prGetAlign(align, index)
}
\arguments{
\item{index}{The index of the align parameter of interest}
}
\description{
Gets alignment
}
\concept{hidden helper functions for}
\keyword{internal}
