% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlTable_helpers_getRowlabelPos.R
\name{prGetRowlabelPos}
\alias{prGetRowlabelPos}
\title{Gets the rowlabel position}
\usage{
prGetRowlabelPos(cgroup = NULL, pos.rowlabel, header = NULL)
}
\arguments{
\item{cgroup}{A vector, matrix or list of character strings defining major column header. The default
is to have none. These elements are also known as \emph{column spanners}. If you want a column \emph{not}
to have a spanner then put that column as "". If you pass cgroup and \code{n.crgroup} as
matrices you can have column spanners for several rows. See cgroup section below for details.}

\item{header}{A vector of character strings specifying column
header, defaulting to \code{\link[base:colnames]{colnames(x)}}}
}
\value{
\code{integer} Returns the position within the header rows
to print the \code{rowlabel} argument
}
\description{
Gets the rowlabel position
}
\seealso{
Other hidden helper functions for htmlTable: 
\code{\link{prAddCells}()},
\code{\link{prAddSemicolon2StrEnd}()},
\code{\link{prEscapeHtml}()},
\code{\link{prGetCgroupHeader}()},
\code{\link{prGetStyle}()},
\code{\link{prPrepInputMatrixDimensions}()},
\code{\link{prPrepareAlign}()},
\code{\link{prPrepareCgroup}()},
\code{\link{prTblNo}()}
}
\concept{hidden helper functions for htmlTable}
\keyword{internal}
