% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveTable.R
\name{prGetScriptString}
\alias{prGetScriptString}
\title{Gets a string with all the scripts merged into one script tag}
\usage{
prGetScriptString(x)
}
\arguments{
\item{x}{An interactiveTable}
}
\value{
string
}
\description{
Each element has its own script tags; otherwise an error will cause
all the scripts to fail.
}
\keyword{internal}
