% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlTable_helpers_getStyle.R
\name{prGetStyle}
\alias{prGetStyle}
\title{Gets the CSS style element}
\usage{
prGetStyle(...)
}
\arguments{
\item{...}{Styles can be provided as \code{vector}, \verb{named vector}, or \code{string}.
If you provide a name, e.g. \code{background: blue}, \code{align="center"},
the function will convert the \code{align} into proper \code{align: center}.}
}
\value{
\code{string} Returns the codes merged into one string with
correct CSS ; and : structure.
}
\description{
A function for checking, merging, and more
with a variety of different style formats.
}
\seealso{
Other hidden helper functions for htmlTable: 
\code{\link{prAddCells}()},
\code{\link{prAddEmptySpacerCell}()},
\code{\link{prAddSemicolon2StrEnd}()},
\code{\link{prEscapeHtml}()},
\code{\link{prGetCgroupHeader}()},
\code{\link{prGetRowlabelPos}()},
\code{\link{prPrepInputMatrixDimensions}()},
\code{\link{prPrepareAlign}()},
\code{\link{prPrepareCgroup}()},
\code{\link{prTblNo}()}
}
\concept{hidden helper functions for htmlTable}
\keyword{internal}
