% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlTable.R
\name{prIsNotebook}
\alias{prIsNotebook}
\title{Detects if the call is made from within an RStudio Rmd file or a file
with the html_notebook output set.}
\usage{
prIsNotebook()
}
\description{
Detects if the call is made from within an RStudio Rmd file or a file
with the html_notebook output set.
}
\keyword{internal}
