% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prPrepareCellStyles}
\alias{prPrepareCellStyles}
\title{Prepares the cell style}
\usage{
prPrepareCellStyles(x, css.cell, rnames, header)
}
\arguments{
  \item{x}{The matrix/data.frame with the data}

  \item{css.cell}{The css.cell element allows you to add
  any possible CSS style to your table cells. If you
  provide a vector the vector it is assummed that the
  styles should be repeated throughout the columns. If you
  provide a matrix of the same size as your \code{x}
  argument. If have \code{ncol(x) + 1} the first row will
  correspond to the rowname style. Correspondingly if the
  size is \code{nrow(x) + 1} it is assummed that the first
  row is the header row.}

  \item{rnames}{Default rownames are generated from
  \code{\link[base]{rownames}(x)}. If you provide
  \code{FALSE} then it will skip the rownames. \emph{Note:}
  For \code{data.frames} if you do
  \code{\link[base]{rownames}(my_dataframe) <- NULL} it
  still has rownames. Thus you need to use \code{FALSE} if
  you want to surpress rownames for \code{data.frames}.}

  \item{header}{A vector of character strings specifying
  column header, defaulting to
  \code{\link[base]{colnames}(x)}}
}
\value{
\code{matrix}
}
\description{
Prepares the cell style
}
\keyword{internal}

