% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prPrepareColors}
\alias{prPrepareColors}
\title{Prepares the alternating colors}
\usage{
prPrepareColors(clr, n, ng)
}
\arguments{
  \item{clr}{The colors}

  \item{n}{The number of rows/columns applicable to the
  color}

  \item{ng}{The n.rgroup argument if applicable}
}
\value{
\code{character} A vector containing hexadecimal colors
}
\description{
Prepares the alternating colors
}
\keyword{internal}

