% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlTable_helpers_prepInputMatrixDimensions.R
\name{prPrepInputMatrixDimensions}
\alias{prPrepInputMatrixDimensions}
\title{Makes sure the input is correct}
\usage{
prPrepInputMatrixDimensions(x, header = NULL)
}
\arguments{
\item{x}{The matrix/data.frame with the data. For the \code{print} and \code{knit_print}
it takes a string of the class \code{htmlTable} as \code{x} argument.}

\item{header}{A vector of character strings specifying column
header, defaulting to \code{\link[base]{colnames}(x)}}
}
\description{
Checks and converts dimensions into something the
\code{\link{htmlTable}} is comfortable with.
}
\seealso{
Other hidden helper functions for htmlTable: 
\code{\link{prAddCells}()},
\code{\link{prAddSemicolon2StrEnd}()},
\code{\link{prEscapeHtml}()},
\code{\link{prGetCgroupHeader}()},
\code{\link{prGetRowlabelPos}()},
\code{\link{prGetStyle}()},
\code{\link{prPrepareAlign}()},
\code{\link{prPrepareCgroup}()},
\code{\link{prTblNo}()}
}
\concept{hidden helper functions for htmlTable}
\keyword{internal}
