% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlTable_helpers_skipRownames.R
\name{prSkipRownames}
\alias{prSkipRownames}
\title{Returns if rownames should be printed for the htmlTable}
\usage{
prSkipRownames(rnames)
}
\arguments{
\item{rnames}{Default row names are generated from \code{\link[base]{rownames}(x)}. If you
provide \code{FALSE} then it will skip the row names. \emph{Note:} For \code{data.frames}
if you do \code{\link[base]{rownames}(my_dataframe) <- NULL} it still has
row names. Thus you need to use \code{FALSE} if you want to
supress row names for \code{data.frames}.}
}
\description{
Returns if rownames should be printed for the htmlTable
}
\keyword{internal}
