% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pvalueFormatter}
\alias{pvalueFormatter}
\title{Deprecated use \code{\link{txtPval}} instead}
\usage{
pvalueFormatter(...)
}
\arguments{
  \item{...}{Currently only used for generating warnings of
  deprecated call}
}
\description{
Deprecated use \code{\link{txtPval}} instead
}
\examples{
pvalueFormatter(c(0.10234,0.010234, 0.0010234, 0.000010234))
}
\keyword{internal}

