% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{splitLines4Table}
\alias{splitLines4Table}
\title{See \code{\link[=txtMergeLines]{txtMergeLines()}}}
\usage{
splitLines4Table(...)
}
\arguments{
\item{...}{passed onto \code{\link[=txtMergeLines]{txtMergeLines()}}}
}
\description{
See \code{\link[=txtMergeLines]{txtMergeLines()}}
}
\examples{
\dontrun{
# Deprecated function
splitLines4Table("hello", "world")
}
}
\keyword{internal}
