% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tblNoLast}
\alias{tblNoLast}
\title{Gets the last table number}
\usage{
tblNoLast(roman = getOption("table_counter_roman", FALSE))
}
\arguments{
  \item{roman}{Whether or not to use roman numbers instead
  of arabic. Can also be set through
  \code{options(fig_caption_no_roman = TRUE)}}
}
\description{
The function relies on \code{options("table_counter")} in
order to keep track of the last number.
}
\examples{
org_opts <- options(table_counter=1)
tblNoLast()
options(org_opts)
}
\seealso{
Other table functions: \code{\link{htmlTable}},
\code{\link{htmlTable.default}},
\code{\link{print.htmlTable}}; \code{\link{tblNoNext}}
}

