% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlTable_style_handlers.R
\name{hasHtmlTableStyle}
\alias{hasHtmlTableStyle}
\title{Check if object has a style set to it}
\usage{
hasHtmlTableStyle(x, style_name)
}
\arguments{
\item{x}{The object intended for \code{\link[=htmlTable]{htmlTable()}}.}

\item{style_name}{A string that contains the style name.}
}
\value{
\code{logical} \code{TRUE} if the attribute and style is not \code{NULL}
}
\description{
If the attribute \code{htmlTable.style} is set it will check if
the \code{style_name} exists and return a \code{logical}.
}
\examples{
library(magrittr)

mx <- matrix(1:4, ncol = 2)
colnames(mx) <- LETTERS[1:2]
mx \%>\%
  addHtmlTableStyle(align = "l|r") \%>\%
  hasHtmlTableStyle("align")
}
\seealso{
Other htmlTableStyle: 
\code{\link{addHtmlTableStyle}()},
\code{\link{highlightRow}()}
}
\concept{htmlTableStyle}
