% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveTable.R
\name{interactiveTable}
\alias{interactiveTable}
\alias{interactiveTable.htmlTable}
\alias{knit_print.interactiveTable}
\alias{print.interactiveTable}
\title{An interactive table that allows you to limit the size of boxes}
\usage{
interactiveTable(x, ..., txt.maxlen = 20, minimized.columns)

\method{interactiveTable}{htmlTable}(tbl, txt.maxlen = 20, minimized.columns)

\method{knit_print}{interactiveTable}(x, ...)

\method{print}{interactiveTable}(x, useViewer, ...)
}
\arguments{
\item{x}{The interactive table that is to be printed}

\item{...}{The exact same parameters as \code{\link{htmlTable}} uses}

\item{txt.maxlen}{The maximum length of a text}

\item{minimized.columns}{Notifies if any particular columns should be collapsed from start}

\item{tbl}{An htmlTable object can be directly passed into the function}

\item{useViewer}{If you are using RStudio there is a viewer thar can render
the table within that is envoced if in \code{\link[base]{interactive}} mode.
Set this to \code{FALSE} if you want to remove that  functionality. You can
also force the function to call a specific viewer by setting this to a
viewer function, e.g. \code{useViewer = utils::browseUrl} if you want to
override the default RStudio viewer. Another option that does the same is to
set the \code{options(viewer=utils::browseUrl)} and it will default to that
particular viewer (this is how RStudio decides on a viewer).
\emph{Note:} If you want to force all output to go through the
\code{\link[base]{cat}()} the set \code{\link[base]{options}(htmlTable.cat = TRUE)}.}
}
\value{
An htmlTable with a javascript attribute containing the code that is then printed
}
\description{
This function wraps the htmlTable and adds JavaScript code for toggling the amount
of text shown in any particular cell.
}

