% Generated by roxygen2 (4.0.2): do not edit by hand
\name{outputInt}
\alias{outputInt}
\title{Deprecated use \code{\link{txtInt}} instead.}
\usage{
outputInt(...)
}
\arguments{
  \item{...}{Passed to \code{\link{txtInt}}}
}
\description{
Deprecated use \code{\link{txtInt}} instead.
}
\examples{
outputInt(123456)
}
\keyword{internal}

