% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/htmlTable_helpers.R
\name{prAddSemicolon2StrEnd}
\alias{prAddSemicolon2StrEnd}
\title{Add a ; at the end}
\usage{
prAddSemicolon2StrEnd(my_str)
}
\arguments{
  \item{my_str}{The string that is to be processed}
}
\value{
\code{string}
}
\description{
The CSS expects a semicolon at the end of each argument
this function just adds a semicolong if none is given and
remove multiple semicolon if such exist
}
\seealso{
Other hidden helper functions for \code{\link{htmlTable}}:
\code{\link{prAddCells}}; \code{\link{prGetCgroupHeader}};
\code{\link{prGetRowlabelPos}}; \code{\link{prGetStyle}};
\code{\link{prPrepareAlign}};
\code{\link{prPrepareCgroup}}; \code{\link{prTblNo}}
}
\keyword{internal}

