% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlTable_helpers.R
\name{prAttr4RgroupAdd}
\alias{prAttr4RgroupAdd}
\title{Get the add attribute element}
\usage{
prAttr4RgroupAdd(rgroup, rgroup_iterator, no_cols)
}
\arguments{
\item{rgroup}{A vector of character strings containing headings for row groups.
\code{n.rgroup} must be present when \code{rgroup} is given. See
 detailed description in section below.}

\item{rgroup_iterator}{The rgroup number of interest}

\item{no_cols}{The \code{ncol(x)} of the core htmlTable x argument}
}
\description{
Gets the add element attribute if it exists. If non-existant it will
return NULL.
}
\keyword{internal}
