% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyHtmlTable_helpers_bindDataListIntoColumns.r
\name{prBindDataListIntoColumns}
\alias{prBindDataListIntoColumns}
\title{Merge columns into a tibble}
\usage{
prBindDataListIntoColumns(dataList)
}
\description{
Almost the same as \code{\link[tibble]{tibble}} but it solves the issue
with some of the arguments being columns and some just being vectors.
}
