% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlTable_helpers_getAlign.R
\name{prGetAlign}
\alias{prGetAlign}
\title{Gets alignment}
\usage{
prGetAlign(
  align,
  index,
  style_list = NULL,
  spacerCell = FALSE,
  followed_by_spacer_cell = FALSE,
  previous_was_spacer_cell = FALSE
)
}
\arguments{
\item{align}{A character strings specifying column alignments, defaulting to \code{'c'}
to center. Valid chars for alignments are l = left, c = center and r = right. You can also specify
\code{align='c|c'} and other LaTeX tabular formatting. If you want to set the alignment of the
rownames this string needst to be \code{ncol(x) + 1}, otherwise it automatically
pads the string with a left alignment for the rownames.}

\item{index}{The index of the align parameter of interest}
}
\description{
Gets alignment
}
\concept{hidden helper functions for}
\keyword{internal}
