% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlTable_helpers.R
\name{prPrepareColors}
\alias{prPrepareColors}
\title{Prepares the alternating colors}
\usage{
prPrepareColors(clr, n, ng, gtxt)
}
\arguments{
\item{clr}{The colors}

\item{n}{The number of rows/columns applicable to the color}

\item{ng}{The n.rgroup/n.cgroup argument if applicable}

\item{gtxt}{The rgroup/cgroup texts}
}
\value{
\code{character} A vector containing hexadecimal colors
}
\description{
Prepares the alternating colors
}
\keyword{internal}
