% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlTable_helpers_prepareCss.R
\name{prPrepareCss}
\alias{prPrepareCss}
\title{Prepares the cell style}
\usage{
prPrepareCss(
  x,
  css,
  rnames,
  header = NULL,
  name = deparse(substitute(css)),
  style_list = NULL
)
}
\arguments{
\item{x}{The matrix/data.frame with the data. For the \code{print} and \code{knit_print}
it takes a string of the class \code{htmlTable} as \code{x} argument.}

\item{css}{The CSS styles that are to be converted into
a matrix.}

\item{rnames}{Default row names are generated from \code{\link[base]{rownames}(x)}. If you
provide \code{FALSE} then it will skip the row names. \emph{Note:} For \code{data.frames}
if you do \code{\link[base]{rownames}(my_dataframe) <- NULL} it still has
row names. Thus you need to use \code{FALSE} if you want to
supress row names for \code{data.frames}.}

\item{header}{A vector of character strings specifying column
header, defaulting to \code{\link[base]{colnames}(x)}}

\item{name}{The name of the CSS style that is prepared}
}
\value{
\code{matrix}
}
\description{
Prepares the cell style
}
\keyword{internal}
