% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prSkipRownames}
\alias{prSkipRownames}
\title{Returns if rownames should be printed for the htmlTable}
\usage{
prSkipRownames(rnames)
}
\arguments{
  \item{rnames}{Default rownames are generated from
  \code{\link[base]{rownames}(x)}. If you provide
  \code{FALSE} then it will skip the rownames. \emph{Note:}
  For \code{data.frames} if you do
  \code{\link[base]{rownames}(my_dataframe) <- NULL} it
  still has rownames. Thus you need to use \code{FALSE} if
  you want to surpress rownames for \code{data.frames}.}
}
\description{
Returns if rownames should be printed for the htmlTable
}
\keyword{internal}

