% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlTable_helpers.R
\name{prTblNo}
\alias{prTblNo}
\title{Gets the table counter string}
\usage{
prTblNo(caption)
}
\arguments{
\item{The}{caption}
}
\value{
\code{string} Returns a string formatted according to
 the table_counter_str and table_counter_roman. The number is
 decided by the table_counter variable
}
\description{
Returns the string used for htmlTable to number the different tables.
Uses options \code{table_counter}, \code{table_counter_str},
and \code{table_counter_roman} to produce the final string. You
can set each option by simply calling \code{options()}.
}
\seealso{
Other hidden helper functions for htmlTable: \code{\link{prAddCells}},
  \code{\link{prAddSemicolon2StrEnd}},
  \code{\link{prEscapeHtml}},
  \code{\link{prGetCgroupHeader}},
  \code{\link{prGetRowlabelPos}}, \code{\link{prGetStyle}},
  \code{\link{prPrepareAlign}},
  \code{\link{prPrepareCgroup}}
}
\concept{hidden helper functions for htmlTable}
\keyword{internal}
