% Generated by roxygen2 (4.0.2): do not edit by hand
\name{splitLines4Table}
\alias{splitLines4Table}
\title{See \code{\link{txtMergeLines}}}
\usage{
splitLines4Table(...)
}
\arguments{
  \item{...}{passed onto \code{\link{txtMergeLines}}}
}
\description{
See \code{\link{txtMergeLines}}
}
\examples{
splitLines4Table("hello", "world")
}
\keyword{internal}

