% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblNo.R
\name{tblNoLast}
\alias{tblNoLast}
\title{Gets the last table number}
\usage{
tblNoLast(roman = getOption("table_counter_roman", FALSE))
}
\arguments{
\item{roman}{Whether or not to use roman numbers instead
of arabic. Can also be set through \code{options(table_caption_no_roman = TRUE)}}
}
\description{
The function relies on \code{options("table_counter")}
in order to keep track of the last number.
}
\examples{
org_opts <- options(table_counter=1)
tblNoLast()
options(org_opts)
}
\seealso{
Other table functions: 
\code{\link{htmlTable}},
\code{\link{tblNoNext}()}
}
\concept{table functions}
