% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyHtmlTable.R
\name{tidyHtmlTable}
\alias{tidyHtmlTable}
\title{Generate an htmlTable using a ggplot2-like interface}
\usage{
tidyHtmlTable(x, value = "value", header = "header", rnames = "rnames",
  rgroup = NULL, hidden_rgroup = NULL, cgroup1 = NULL, cgroup2 = NULL,
  tspanner = NULL, hidden_tspanner = NULL, ...)
}
\arguments{
\item{x}{Tidy data used to build the \code{htmlTable}}

\item{value}{The column containing values filling individual cells of the
output \code{htmlTable}}

\item{header}{The column in \code{x} specifying column headings}

\item{rnames}{The column in \code{x} specifying row names}

\item{rgroup}{The column in \code{x} specifying row groups}

\item{hidden_rgroup}{rgroups that will be hidden.}

\item{cgroup1}{The column in \code{x} specifying the inner most column
groups}

\item{cgroup2}{The column in \code{x} specifying the outer most column
groups}

\item{tspanner}{The column in \code{x} specifying tspanner groups}

\item{hidden_tspanner}{tspanners that will be hidden.}

\item{...}{Additional arguments that will be passed to the inner
\code{htmlTable} function}
}
\value{
Returns html code that will build a pretty table
}
\description{
Builds an \code{htmlTable} by mapping columns from the input data, \code{x},
to elements of output \code{htmlTable} (e.g. rnames, headers, etc.)
}
\examples{
\dontrun{
library(tidyverse)
mtcars \%>\%
    rownames_to_column \%>\%
    select(rowname, cyl, gear, hp, mpg, qsec) \%>\%
    gather(per_metric, value, hp, mpg, qsec) \%>\%
    group_by(cyl, gear, per_metric) \%>\%
    summarise(Mean = round(mean(value), 1),
              SD = round(sd(value), 1),
              Min = round(min(value), 1),
              Max = round(max(value), 1)) \%>\%
     gather(summary_stat, value, Mean, SD, Min, Max) \%>\%
     ungroup \%>\%
     mutate(gear = paste(gear, "Gears"),
            cyl = paste(cyl, "Cylinders")) \%>\%
     tidyHtmlTable(header = "gear",
                  cgroup1 = "cyl",
                  cell_value = "value",
                  rnames = "summary_stat",
                  rgroup = "per_metric")
}
}
\seealso{
\code{\link{htmlTable}}
}
