% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyHtmlTable.R
\name{tidyHtmlTable}
\alias{tidyHtmlTable}
\title{Generate an htmlTable using tidy data as input}
\usage{
tidyHtmlTable(
  x,
  value,
  header,
  rnames,
  rgroup,
  hidden_rgroup,
  cgroup,
  tspanner,
  hidden_tspanner,
  skip_removal_warning = getOption("htmlTable.skip_removal_warning", FALSE),
  table_fn = htmlTable,
  ...
)
}
\arguments{
\item{x}{Tidy data used to build the \code{htmlTable}}

\item{value}{The column containing values filling individual cells of the
output \code{htmlTable}. Defaults to "value" as used by \code{\link[tidyr]{pivot_longer}}.}

\item{header}{The column in \code{x} specifying column headings}

\item{rnames}{The column in \code{x} specifying row names. Defaults to "name" as used by
\code{\link[tidyr]{pivot_longer}}.}

\item{rgroup}{The column in \code{x} specifying row groups}

\item{hidden_rgroup}{\code{strings} with rgroup values that will be hidden  (the values will
still be there but thhe spanner will be set to "" and thus ignored byt \code{\link{htmlTable}}).}

\item{cgroup}{The column or columns in \code{x} specifying the column groups}

\item{tspanner}{The column in \code{x} specifying tspanner groups}

\item{hidden_tspanner}{\code{strings} with tspanner values that will be hidden (the values will
still be there but thhe spanner will be set to "" and thus ignored byt \code{\link{htmlTable}}).}

\item{skip_removal_warning}{\code{boolean} supress warning message when removing NA columns.}

\item{table_fn}{The table function that should receive the input, defaults to \code{\link{htmlTable}}
but you can provide any function that uses the same input formatting. This package was inspired
by the \code{\link[Hmisc]{latex}} function.}

\item{...}{Additional arguments that will be passed to the inner
\code{\link{htmlTable}} function}
}
\value{
Returns html code that will build a pretty table
}
\description{
Builds an \code{htmlTable} by mapping columns from the input data, \code{x},
to elements of an output \code{htmlTable} (e.g. rnames, header, etc.). This
provides a ggplot2-like interface you can pivot rows/columns as required. The
typical use case is when you are using \code{dplyr} together with the
\code{tidyverse} data processing functions, see \code{vignette("tidyHtmlTable")}.
}
\section{Column-mapping parameters}{


  The \code{tidyHtmlTable} function is designed to work like ggplot2 in that
  columns from \code{x} are mapped to specific parameters from the
  \code{htmlTable} function. At minimum, \code{x} must contain the names
  of columns mapping to \code{rnames}, \code{header}, and \code{rnames}.
  \code{header} and \code{rnames} retain the same meaning as in the
  htmlTable function. \code{value} contains the individual values that will
  be used to fill each cell within the output \code{htmlTable}.

  A full list of parameters from \code{htmlTable} which may be mapped to
  columns within \code{x} include:

  \itemize{
    \item \code{value}
    \item \code{header}
    \item \code{rnames}
    \item \code{rgroup}
    \item \code{cgroup}
    \item \code{tspanner}
  }

  Also note that the coordinates of each \code{value} within \code{x} must be
  unambiguously mapped to a position within the output \code{htmlTable}.
  Therefore, the each row-wise combination the variables specified above
  contained in \code{x} must be unique.
}

\section{Sorting}{


  Sorting of rows is as of version 2.0 skipped as we may have situations with
  repeating inputs and this can easily be performed pre-function by calling
  \code{\link[dplyr]{arrange}()} prior to \code{tidyHtmlTable}.

  Columns are sorted by \code{arrange(cgroup, header)} where cgroup will be
  expanded to the columns of the cgroup argument, e.g. \code{cgroup = c(a, b), header = c}
  will become \code{arrange(a, b, c)}. If you want to sort in non-alphabetic order
  you can provide a \code{factor} variable and that information will be retained.
}

\section{Hidden values}{


  \code{htmlTable} Allows for some values within \code{rgroup},
  \code{cgroup}, etc. to be specified as \code{""}. The following parameters
  allow for specific values to be treated as if they were a string of length
  zero in the \code{htmlTable} function.

  \itemize{
    \item \code{hidden_rgroup}
    \item \code{hidden_tspanner}
  }
}

\section{Simple tibble output}{


  The tibble discourages the use of rownames. There is therefore a convenience
  option for \code{tidyHtmlTable} where you can use the function just as you
  would with htmlTable where \code{rnames} is populated with "names" column or
  the \code{rnames} argument provided using tidyselect syntax.
}

\section{Additional dependencies}{


   In order to run this function you also must have \code{\link[dplyr]{dplyr}},
   \code{\link[tidyr]{tidyr}}, \code{\link[tidyselect]{tidyselect}} and
   \code{\link[purrr]{purrr}} packages installed. These have been removed due to
   the additional 20 Mb that these dependencies added (issue #47).
}

\examples{
library(tibble)
library(dplyr)
library(tidyr)

mtcars \%>\%
  rownames_to_column() \%>\%
  select(rowname, cyl, gear, hp, mpg, qsec) \%>\%
  pivot_longer(names_to = "per_metric",
               cols = c(hp, mpg, qsec)) \%>\%
  group_by(cyl, gear, per_metric) \%>\%
  summarise(
    Mean = round(mean(value), 1),
    SD = round(sd(value), 1),
    Min = round(min(value), 1),
    Max = round(max(value), 1)
  ) \%>\%
  pivot_longer(names_to = "summary_stat",
               cols = c(Mean, SD, Min, Max)) \%>\%
  ungroup() \%>\%
  mutate(
    gear = paste(gear, "Gears"),
    cyl = paste(cyl, "Cylinders")
  ) \%>\%
  addHtmlTableStyle(align = "r") \%>\%
  tidyHtmlTable(
    header = gear,
    cgroup = cyl,
    rnames = summary_stat,
    rgroup = per_metric,
    skip_removal_warning = TRUE)
}
\seealso{
\code{\link{htmlTable}}
}
