% Generated by roxygen2 (4.0.2): do not edit by hand
\name{txtMergeLines}
\alias{txtMergeLines}
\title{A merges lines while preserving the line break for html/LaTeX}
\usage{
txtMergeLines(..., html = TRUE)
}
\arguments{
  \item{...}{The lines that you want to be joined}

  \item{html}{If HTML compatible output should be used. If
  \code{FALSE} it outputs LaTeX formatting}
}
\value{
string
}
\description{
This function helps you to do a multiline table header in
both html and in LaTeX. In html this isn't that tricky, you
just use the <br /> command but in LaTeX I often find
myself writing vbox/hbox stuff and therefore I've created
this simple helper function
}
\examples{
txtMergeLines("hello", "world")
txtMergeLines("hello", "world", html=FALSE)
txtMergeLines("hello", "world", list("A list", "is OK"))
}
\seealso{
Other text formatters: \code{\link{txtPval}};
\code{\link{txtRound}}
}

