% Generated by roxygen2 (4.0.2): do not edit by hand
\name{txtRound}
\alias{txtRound}
\title{A convenient rounding function}
\usage{
txtRound(x, digits, excl.cols, excl.rows)
}
\arguments{
  \item{x}{The data.frame/matrix to be rounded}

  \item{digits}{The number of digits to round each element
  to. If you provide a vector each element for
  corresponding columns.}

  \item{excl.cols}{Rows to exclude from the rounding
  procedure. This can be either a number or regular
  expression.}

  \item{excl.rows}{Columns to exclude from the rounding
  procedure. This can be either a number or regular
  expression.}
}
\value{
\code{matrix/data.frame}
}
\description{
A convenient rounding function
}
\examples{
mx <- matrix(c(1, 1.11, 1.25,
               2.50, 2.55, 2.45,
               3.2313, 3, pi),
             ncol = 3, byrow=TRUE)
txtRound(mx, 1)
}
\seealso{
Other text formatters: \code{\link{txtMergeLines}};
\code{\link{txtPval}}
}

